Сorrected edits to Storybook Reorganisation # 31 pullrequest
