+++ stories/AntDefault/Data Entry/Rate Mobile/index.js
