+++ stories/AntDefault/Data Entry/Slider Mobile/index.js
